luci-app-ssr-plus:Code optimization and Fix Bug (#4253)
* luci-app-ssr-plus:Code optimization and Fix Bug

* luci-app-ssr-plus:Adjust reference data location

+++ package/lean/luci-app-ssr-plus/root/usr/share/shadowsocksr/chinaipset.sh
@@ -1,5 +1,9 @@
+#!/bin/sh
+LOCK_FILE="/var/lock/ssr-chinaipset.lock"
+[ -f "$LOCK_FILE" ] && exit 2
+touch "$LOCK_FILE"
 echo "create china hash:net family inet hashsize 1024 maxelem 65536" > /tmp/china.ipset
 awk '!/^$/&&!/^#/{printf("add china %s'" "'\n",$0)}' /etc/china_ssr.txt >> /tmp/china.ipset
 ipset -! flush china
 ipset -! restore < /tmp/china.ipset 2>/dev/null
-rm -f /tmp/china.ipset
+rm -f /tmp/china.ipset $LOCK_FILE
